bookings can now only be fetched and cancelled by their owner users

+++ src/graphql/resolvers/booking.ts
@@ -12,7 +12,7 @@ module.exports = {
       throw new Error('Unauthenticated')
     }
     try {
-      const bookings: Array<IBookingDocument> = await Booking.find()
+      const bookings: Array<IBookingDocument> = await Booking.find({ user: req.userId })
       return bookings.map(
         (booking: IBookingDocument): IBookingDocument['_doc'] => {
           return standarizeBooking(booking)
@@ -44,6 +44,9 @@ module.exports = {
     }
     try {
       const booking = await Booking.findById(args.bookingId).populate('event')
+      if (req.userId !== booking.user) {
+        throw new Error('User not deleting its own booking')
+      }
       const event: IEventDocument['_doc'] = standarizeEvent(booking.event)
       await Booking.deleteOne({ _id: args.bookingId })
       return event
